pyproject.toml
[tool.poetry]
name = "django_lowcoder"
version = "0.1.0"
description = "Web application to generate crud web applications out ofexcel spreadsheets"
authors = ["Henry Jobst <github@jobst-berlin.de>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.4"
django-crispy-forms = "^1.14.0"
crispy-bootstrap5 = "^0.7"
cookiecutter = "^2.1.1"
fs = "^2.4.16"
django-braces = "^1.15.0"
django-debug-toolbar = "^3.8.1"
django-extensions = "^3.2.1"
pipx = {version = "^1.1.0"}
black = {version = "^22.12.0"}
types-python-slugify = {version = "^7.0.0"}
pytest = "^7.2.0"
exceptiongroup = "^1.1.0"
django-environ = "^0.9.0"
django-htmx = "^1.13.0"
pandas = "^1.5.2"

[tool.poetry.dev-dependencies]
tox = {version = "^4.0.16"}
pytest = {version = "^7.2.0"}
mypy = {version = "^0.991"}
exceptiongroup = {version = "^1.1.0"}
pytest-django = "^4.5.2"
django-stubs="^1.13.1"
pytest-sugar="^0.9.6"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--ignore output --ds=django_lowcoder.config.settings.test --reuse-db"
python_files = "tests.py, test_*.py"
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]

[tool.mypy]
exclude = [
    "output/",
]

[tool.tox]
legacy_tox_ini = """
    [tox]
    skipsdist=true
    env_list = py311,format
    minversion = 4.0.16

    [testenv]
    description = run the tests with pytest
    package = wheel
    wheel_build_env = .pkg
    deps = -rrequirements.txt
    allowlist_externals: pytest
    commands = pytest {tty:--color=yes} {posargs}

    [testenv:format]
    description = install black in a virtual environment and invoke it on the current folder
    deps = black
    skip_install = true
    commands = black .

    [testenv:type]
    deps = mypy
    commands = mypy .
"""
